requirements.txt
numpy==2.2.0
Pillow==11.0.0
torch==2.4.1
